stuf was done

+++ build.py
@@ -5,7 +5,8 @@ import subprocess
 
 v = input("Enter the version: ")
 subprocess.call("pyinstaller main.py -F -w --clean -p ./assets/scripts")
-shutil.rmtree(f"./builds/{v}", ignore_errors=True)
+if os.path.isdir(f"./builds/{v}"):
+    shutil.rmtree(f"./builds/{v}", ignore_errors=True)
 os.mkdir(f"./builds/{v}")
 shutil.copytree("./assets", f"./builds/{v}/assets")
 shutil.move("./dist/main.exe", f"./builds/{v}")
